zna-app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
google-generativeai>=0.7.0
reportlab>=4.2.0
tenacity>=8.3.0
slowapi>=0.1.9
python-multipart>=0.0.9
